requirements.txt
anthropic>=0.40.0
openai>=1.50.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
